misc. changes & fixes
git-svn-id: https://jmoney.svn.sf.net/svnroot/jmoney/trunk@278 b97006a3-7e3d-0410-986d-ea81e883d316

+++ net.sf.jmoney/src/net/sf/jmoney/model2/Entry.java
@@ -323,7 +323,7 @@ public class Entry extends ExtendableObject {
 		amount = anAmount;
 		
 		// Notify the change manager.
-		processPropertyChange(EntryInfo.getAmountAccessor(), new Double(oldAmount), new Double(amount));
+		processPropertyChange(EntryInfo.getAmountAccessor(), new Long(oldAmount), new Long(amount));
 	}
 	
 	/**
